tests/wpt/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.round.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.create2.round</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.create2.round</h1>
<p class="desc">createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)</p>


<script>
var t = async_test("createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata1 = ctx.createImageData(10.01, 10.99);
var imgdata2 = ctx.getImageData(0, 0, 10.01, 10.99);
_assertSame(imgdata1.width, imgdata2.width, "imgdata1.width", "imgdata2.width");
_assertSame(imgdata1.height, imgdata2.height, "imgdata1.height", "imgdata2.height");
t.done();

});
</script>
