LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.round-expected.txt

2d.imageData.create2.round

createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)


PASS createImageData(w, h) is rounded the same as getImageData(0, 0, w, h) 
